requirements.txt
arrow==1.2.3
boto3==1.26.164
python-telegram-bot==20
requests==2.31.0